Update compiler options and fix initialization of span object in main.cpp

+++ ArgumentParser/main.cpp
@@ -1,10 +1,11 @@
+#include <cstddef>
 #include <iostream>
 #include <span>
 
 #include "argument_parser.hpp"
 
 int main(const int argc, const char* const* const argv) {
-  std::span<const char* const> args(argv, argc);
+  std::span<const char* const> args(argv, static_cast<size_t>(argc));
 
   lrd::ArgumentParser parser{args};
   std::cout << parser << '\n';
